feat(espower): disambiguate between function calls and async/yield expressions
In order to fully disambiguate the contextual keywords `async`
and `yield`, we need to pass along some addtional context to
power-assert-formatter. Specifically whether the enclosing function
is async or a generator.

This adds two new properties to the context passed to `assert._expr()`

 - `async: true` if the function enclosing the assertion is an async function
 - `generator: true` if the function enclosing the assertion is a generator

In both cases, we do not set a property if the enclosing function is not
a generator/async (i.e. we never pass `async:false`). This is primarily
to avoid having to rewrite a bunch of existing unit tests.